requirements.txt
scikit-image
numpy >= 1.14.5
Pillow >= 5.3.0
